--- plugins/downloader/actions.js
@@ -1,11 +1,0 @@
-const CHANNEL = "downloader";
-const ACTIONS = {
-	ERROR: "error",
-	METADATA: "metadata",
-	PROGRESS: "progress",
-};
-
-module.exports = {
-	CHANNEL: CHANNEL,
-	ACTIONS: ACTIONS,
-};
